devtools/server/tests/mochitest/test_device.html

<!DOCTYPE HTML>
<html>
<!--
Bug 895360 - [app manager] Device meta data actor
-->
<head>
  <meta charset="utf-8">
  <title>Mozilla Bug</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script>
"use strict";

window.onload = function() {
  const {require} = ChromeUtils.import("resource://devtools/shared/Loader.jsm");
  const {DebuggerClient} = require("devtools/shared/client/debugger-client");
  const {DebuggerServer} = require("devtools/server/main");
  const Services = require("Services");

  SimpleTest.waitForExplicitFinish();

  DebuggerServer.init();
  DebuggerServer.registerAllActors();

  const client = new DebuggerClient(DebuggerServer.connectPipe());
  client.connect().then(function onConnect() {
    return client.mainRoot.getFront("device");
  }).then(function(d) {
    let desc;
    const appInfo = Services.appinfo;
    const utils = window.windowUtils;

    const localDesc = {
      appid: appInfo.ID,
      vendor: appInfo.vendor,
      name: appInfo.name,
      version: appInfo.version,
      appbuildid: appInfo.appBuildID,
      platformbuildid: appInfo.platformBuildID,
      platformversion: appInfo.platformVersion,
      geckobuildid: appInfo.platformBuildID,
      geckoversion: appInfo.platformVersion,
      useragent: window.navigator.userAgent,
      locale: Services.locale.appLocaleAsLangTag,
      os: appInfo.OS,
      processor: appInfo.XPCOMABI.split("-")[0],
      compiler: appInfo.XPCOMABI.split("-")[1],
      dpi: utils.displayDPI,
      width: window.screen.width,
      height: window.screen.height,
    };

    function checkValues() {
      for (const key in localDesc) {
        is(desc[key], localDesc[key], "valid field (" + key + ")");
      }

      const currProfD = Services.dirsvc.get("ProfD", Ci.nsIFile);
      const profileDir = currProfD.path;
      ok(profileDir.includes(desc.profile.length > 0 && desc.profile),
         "valid profile name");

      client.close().then(() => {
        DebuggerServer.destroy();
        SimpleTest.finish();
      });
    }

    d.getDescription().then(function(v) {
      desc = v;
    }).then(checkValues);
  });
};
</script>
</pre>
</body>
</html>
